requirements.txt
pyaudio
google-genai
google-generativeai
pillow
requests
beautifulsoup4
playwright
pyautogui
pyperclip
opencv-python
numpy
mss
psutil
send2trash
comtypes
pycaw
win10toast
duckduckgo-search
youtube-transcript-api
python-dotenv